requirements.txt
gradio==4.44.0
pillow==10.4.0
transformers==4.44.2
python-dotenv==1.0.1